requirements.txt
flask
faiss-cpu
pandas
numpy
gdown
sentence-transformers